backend/pyproject.toml
[project]
name = "hng-vibes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.4",
    "httpx>=0.28.1", 
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "rq>=2.6.0",
    "sqlalchemy>=2.0.43",
    "python-dotenv==1.0.0",
    "spotipy==2.23.0",
    "slack_bolt",
    "uvicorn[standard]==0.24.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pylint>=3.3.8",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]
